+++ src/ServerApi/Infrastructures/Adnc.Infra.Gateway/Properties/launchSettings.json
@@ -12,7 +12,9 @@
       "commandName": "IISExpress",
       "launchBrowser": true,
       "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
+        "SKYWALKING__SERVICENAME": "adnc.gateway.dev",
+        "ASPNETCORE_ENVIRONMENT": "Development",
+        "ASPNETCORE_HOSTINGSTARTUPASSEMBLIES": "SkyAPM.Agent.AspNetCore"
       }
     },
     "Adnc.Infra.Gateway": {
