+++ src/ServerApi/Infrastructures/Adnc.Infra.Gateway/appsettings.development.json
@@ -9,7 +9,7 @@
   "AllowedHosts": "*",
   "CorsHosts": "http://localhost:5001",
   "SkyWalking": {
-    "ServiceName": "adnc.ocelot.dev",
+    "ServiceName": "adnc.gateway.dev",
     "Namespace": "adnc",
     "HeaderVersions": [
       "sw8"
@@ -17,7 +17,7 @@
     "Sampling": {
       "SamplePer3Secs": -1,
       "Percentage": -1.0,
-      "IgnorePaths": [ "/*/health-24b01005-a76a-4b3b-8fb1-5e0f2e9564fb", "/*/index.html", "/**/swagger.json" ]
+      "IgnorePaths": [ "/*/health-24b01005-a76a-4b3b-8fb1-5e0f2e9564fb", "http://**/appsettings", "/**/swagger.json" ]
     },
     "Logging": {
       "Level": "Error",
