+++ src/ServerApi/Services/Adnc.Cus/Adnc.Cus.Application.Contracts/Services/ICustomerAppService.cs
@@ -1,6 +1,7 @@
 ﻿using Adnc.Application.Shared.Dtos;
 using Adnc.Application.Shared.Interceptors;
 using Adnc.Application.Shared.Services;
+using Adnc.Core.Shared.Interceptors;
 using Adnc.Cus.Application.Contracts.Dtos;
 using System.Threading.Tasks;
 
@@ -11,7 +12,17 @@ namespace Adnc.Cus.Application.Contracts.Services
         [OpsLog(LogName = "注册")]
         Task<AppSrvResult<CustomerDto>> RegisterAsync(CustomerRegisterDto inputDto);
 
-        [OpsLog(LogName = "充值")] Task<AppSrvResult<SimpleDto<string>>> RechargeAsync(long id, CustomerRechargeDto inputDto);
+        [OpsLog(LogName = "充值")]
+        [UnitOfWork(SharedToCap = true)]
+        Task<AppSrvResult<SimpleDto<string>>> RechargeAsync(long id, CustomerRechargeDto inputDto);
+
+        [OpsLog(LogName = "处理充值")]
+        [UnitOfWork]
+        Task<AppSrvResult> ProcessRechargingAsync(long transactionLogId, long customerId, decimal amount);
+
+        [OpsLog(LogName = "处理付款")]
+        [UnitOfWork(SharedToCap = true)]
+        Task<AppSrvResult> ProcessPayingAsync(long transactionLogId, long customerId, decimal amount);
 
         Task<AppSrvResult<PageModelDto<CustomerDto>>> GetPagedAsync(CustomerSearchPagedDto search);
     }
