+++ src/ServerApi/Services/Adnc.Maint/Adnc.Maint.Application.Contracts/Services/IDictAppService.cs
@@ -1,5 +1,6 @@
 ﻿using Adnc.Application.Shared.Interceptors;
 using Adnc.Application.Shared.Services;
+using Adnc.Core.Shared.Interceptors;
 using Adnc.Infra.Caching.Interceptor;
 using Adnc.Maint.Application.Contracts.Consts;
 using Adnc.Maint.Application.Contracts.Dtos;
@@ -30,6 +31,7 @@ namespace Adnc.Maint.Application.Contracts.Services
         /// <returns></returns>
         [OpsLog(LogName = "修改字典")]
         [CachingEvict(CacheKey = CachingConsts.DictListCacheKey)]
+        [UnitOfWork]
         Task<AppSrvResult> UpdateAsync(long id, DictUpdationDto input);
 
         /// <summary>
