+++ src/ServerApi/Services/Adnc.Maint/Adnc.Maint.WebApi/Properties/launchSettings.json
@@ -13,6 +13,7 @@
       "launchBrowser": true,
       "environmentVariables": {
         "ASPNETCORE_ENVIRONMENT": "Development",
+        "SKYWALKING__SERVICENAME": "adnc.maint.webapi.dev",
         "ASPNETCORE_HOSTINGSTARTUPASSEMBLIES": "SkyAPM.Agent.AspNetCore"
       }
     },
