+++ src/ServerApi/Services/Adnc.Usr/Adnc.Usr.Application.Contracts/Services/IDeptAppService.cs
@@ -1,5 +1,6 @@
 ﻿using Adnc.Application.Shared.Interceptors;
 using Adnc.Application.Shared.Services;
+using Adnc.Core.Shared.Interceptors;
 using Adnc.Infra.Caching.Interceptor;
 using Adnc.Usr.Application.Contracts.Consts;
 using Adnc.Usr.Application.Contracts.Dtos;
@@ -30,6 +31,7 @@ namespace Adnc.Usr.Application.Contracts.Services
         /// <returns></returns>
         [OpsLog(LogName = "修改部门")]
         [CachingEvict(CacheKeys = new string[] { CachingConsts.DetpListCacheKey, CachingConsts.DetpTreeListCacheKey, CachingConsts.DetpSimpleTreeListCacheKey })]
+        [UnitOfWork]
         Task<AppSrvResult> UpdateAsync(long id, DeptUpdationDto input);
 
         /// <summary>
