+++ src/ServerApi/Services/Adnc.Usr/Adnc.Usr.Application.Contracts/Services/IRoleAppService.cs
@@ -1,6 +1,7 @@
 ﻿using Adnc.Application.Shared.Dtos;
 using Adnc.Application.Shared.Interceptors;
 using Adnc.Application.Shared.Services;
+using Adnc.Core.Shared.Interceptors;
 using Adnc.Infra.Caching.Interceptor;
 using Adnc.Usr.Application.Contracts.Consts;
 using Adnc.Usr.Application.Contracts.Dtos;
@@ -48,6 +49,7 @@ namespace Adnc.Usr.Application.Contracts.Services
         /// <returns></returns>
         [OpsLog(LogName = "设置角色权限")]
         [CachingEvict(CacheKeys = new[] { CachingConsts.MenuRelationCacheKey, CachingConsts.MenuCodesCacheKey })]
+        [UnitOfWork]
         Task<AppSrvResult> SetPermissonsAsync(RoleSetPermissonsDto input);
 
         /// <summary>
