+++ src/ServerApi/Services/Adnc.Usr/Adnc.Usr.WebApi/appsettings.development.json
@@ -80,7 +80,7 @@
     "Sampling": {
       "SamplePer3Secs": -1,
       "Percentage": -1.0,
-      "IgnorePaths": [ "/*/health-24b01005-a76a-4b3b-8fb1-5e0f2e9564fb", "/v1/kv/*/*/*/appsettings", "/*/index.html", "/**/swagger.json" ]
+      "IgnorePaths": [ "/*/health-24b01005-a76a-4b3b-8fb1-5e0f2e9564fb", "http://**/appsettings", "/**/swagger.json" ]
     },
     "Logging": {
       "Level": "Error",
